pyproject.toml
[tool.poetry]
name = "workspace"
version = "1.0.0"
description = "Monorepo with shared brain intelligence"
authors = []

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
pydantic = "^2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
mypy = "^1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
strict = true
